+++ imports/mailgun/server/incoming/router.js
@@ -1,11 +1,10 @@
 import './mailgun-emails';
 
-Router.configureBodyParsers = function() {
-  Router.onBeforeAction(Iron.Router.bodyParser.urlencoded({
-    extended: true,
-    limit: '50mb'
-  }));
-};
+Router.onBeforeAction(Iron.Router.bodyParser.urlencoded({
+  extended: true,
+  limit: '10mb'
+}));
+
 // mailgun 收取邮件
 Router.route('/api/v1/mailgun', {
   where: 'server'